feat(diagram-maker): transparent node bg, local diagram search, font fixes
- add transparentBackground option to BaseNode + sidebar checkbox control
- add search input to filter the local-diagrams list in the import/export modal
- fix font-family cascade bug: diagram-maker's own globals.css Tailwind
  Preflight import was overriding fui-components' custom monobit font rule
- bump undersized text-xs/text-sm controls in the sidebar to the app's
  custom text-small/text-caption scale (monobit needs larger sizing)
- fix line-height on select/input controls so labels aren't pushed down
- split ImportExportModal.tsx into ExportTab/ImportTab/LocalDiagramsTab
