+++ apps/diagram-maker/src/app/globals.css
@@ -1,3 +1,26 @@
 @import 'tailwindcss';
 
 @config '../../tailwind.config.ts';
+
+@layer base {
+  @font-face {
+    font-family: 'ethnocentric';
+    src: url('/fonts/ethnocentric_rg.otf') format('opentype');
+    font-weight: normal;
+    font-style: normal;
+  }
+  @font-face {
+    font-family: 'monobit';
+    src: url('/fonts/monobit.ttf') format('truetype');
+    font-weight: normal;
+    font-style: normal;
+  }
+  html {
+    font-family: 'monobit', system-ui, sans-serif;
+    font-size: 16px;
+  }
+  body {
+    font-size: 1.75rem;
+    letter-spacing: 0;
+  }
+}
